Lab_03/requirements.txt
PyQt5
requests

Flask>=2.3.2
rsa>=4.9

pycryptodome

ecdsa
